+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+- [PR-48](https://github.com/itk-dev/ai-screening/pull/48)
+  Access checks in Twig
 - [PR-46](https://github.com/itk-dev/ai-screening/pull/46)
   2991: Added breadcrumb trails
 - [PR-45](https://github.com/itk-dev/ai-screening/pull/45)